Refactor MIME types and add HTTP 405 status code
Extract MIME type mappings to a dedicated header file for better
organization. Add HTTP 405 Method Not Allowed status code and refactor
status codes array to use constants. Fix file size function to use off_t
type for proper large file support. Improve README formatting.

+++ utils.h
@@ -1,5 +1,6 @@
 #ifndef UTILS_H
 #define UTILS_H
+#include <sys/types.h>
 
 // Returns the file extension (including the dot) of a filename
 // Returns NULL if no extension is found
@@ -11,6 +12,6 @@ const char* get_content_type_using_file_extension(const char* extension);
 
 // Returns the size of a file from a file descriptor
 // Returns -1 if an error occurs
-int get_file_size_from_fd(int open_fd);
+off_t get_file_size_from_fd(int open_fd);
 
 #endif /* UTILS_H */
